+++ ProjectBank/Server.Entities/Applicant/ApplicantDTO.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace ProjectBank.Server
+namespace ProjectBank.Server.Entities
 {
     public record ApplicantCreateDTO([Required] int ProjectId, int UserId);
 }
