--- ProjectBank/Server.Entities/Class1.cs
@@ -1,5 +1,0 @@
-﻿namespace Server.Entities;
-public class Class1
-{
-
-}
