+++ ProjectBank/Server.Entities/Project/Project.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 
-namespace ProjectBank.Server
+namespace ProjectBank.Server.Entities
 {
     public class Project
     {
