+++ ProjectBank/Server.Entities/Project/ProjectDTO.cs
@@ -1,6 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace ProjectBank.Server
+namespace ProjectBank.Server.Entities
 {
     public record ProjectCreateDTO([Required] string Name, string Description, int AuthorId);
     public record ProjectDTO(int Id, [Required] string Name, string Description, int AuthorId) : ProjectCreateDTO(Name, Description, AuthorId);
