+++ ProjectBank/Server.Entities/User/UserDTO.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace ProjectBank.Server
+namespace ProjectBank.Server.Entities
 {
     public record UserCreateDTO([Required] string Name, [Required] bool IsSupervisor);
     public record UserDTO(int Id, [Required] string Name, [Required] bool IsSupervisor);
