+++ ProjectBank/Server.Entities/View/ViewDTO.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace ProjectBank.Server
+namespace ProjectBank.Server.Entities
 {
     public record ViewCreateDTO([Required] int ProjectId, int UserId);
 }
