+++ ProjectBank/Server.Tests/Server.Tests.csproj
@@ -24,6 +24,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Server\ProjectBank.Server.csproj" />
+    <ProjectReference Include="..\Server.Entities\Server.Entities.csproj" />
   </ItemGroup>
 
 </Project>
