+++ ProjectBank/Server/ProjectBankContextFactory.cs
@@ -2,8 +2,9 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Design;
 using Microsoft.Extensions.Configuration;
 using System.IO;
+using ProjectBank.Server.Entities;
 
-namespace ProjectBank.Server
+namespace ProjectBank.Server.Entities
 {
     public class ProjectBankContextFactory : IDesignTimeDbContextFactory<ProjectBankContext>
     {
